add split invariant test. Passed

+++ components/wmtk_components/adaptive_tessellation/wmtk/components/adaptive_tessellation/invariants/RGBSplitInvariant.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
+#include <Eigen/Dense>
 #include <wmtk/attribute/AttributeHandle.hpp>
 #include <wmtk/invariants/Invariant.hpp>
-
 namespace wmtk {
 class RGBSplitInvariant : public Invariant
 {
@@ -13,15 +13,15 @@ class RGBSplitInvariant : public Invariant
 public:
     RGBSplitInvariant(
         const Mesh& m,
-        const TypedAttributeHandle<double>& face_rgb_state_handle,
-        const TypedAttributeHandle<double>& edge_rgb_state_handle);
+        const TypedAttributeHandle<int64_t>& face_rgb_state_handle,
+        const TypedAttributeHandle<int64_t>& edge_rgb_state_handle);
     bool before(const simplex::Simplex& t) const override;
 
 private:
     bool can_refine(
         const Eigen::Vector2<int64_t>& edge_color_level,
         const Eigen::Vector2<int64_t>& my_face_color_level) const;
-    const TypedAttributeHandle<double> m_face_rgb_state_handle;
-    const TypedAttributeHandle<double> m_edge_rgb_state_handle;
+    const TypedAttributeHandle<int64_t> m_face_rgb_state_handle;
+    const TypedAttributeHandle<int64_t> m_edge_rgb_state_handle;
 };
 } // namespace wmtk
